Add /save yourself/ command.
Use the new versions of ector (0.1.7) and file-concept-network (0.1.2).
Still a little problem: some error messages that are in fact warnings.
It is due to the lack of callbacks to treat errors (during load or save).
I have to figure out how to do this in CoffeeScript (more accustomed to Javascript).